쿠폰생성 및 발급

+++ src/main/java/kr/hhplus/be/server/point/Main.java
@@ -2,10 +2,12 @@ package kr.hhplus.be.server.point;
 
 import kr.hhplus.be.server.domain.point.PointEntity;
 
+import java.math.BigDecimal;
+
 public class Main {
     public static void main(String[] args) {
         PointEntity point = new PointEntity();
-        point.setPointTotal(1000l);
+        point.setPointTotal(BigDecimal.valueOf(1000l));
         System.out.println(point.getPointTotal()); //정상 출력되면 Lombok 문제 없음
     }
 }
